We were allocating a new RowDescription message (and all of its FieldDescriptions) for every new statement execution (except for prepared statements). We now recycle these and reuse the same RowDescription cached on the physical connection. Prepared commands clone the description as they need their own, long-lived instance.