@roji 2019-03-21
# Before

``` ini

BenchmarkDotNet=v0.11.3, OS=ubuntu 18.10
Intel Xeon W-2133 CPU 3.60GHz, 1 CPU, 12 logical and 6 physical cores
.NET Core SDK=3.0.100-preview-009812
  [Host]     : .NET Core 3.0.0-preview-27122-01 (CoreCLR 4.6.27121.03, CoreFX 4.7.18.57103), 64bit RyuJIT
  DefaultJob : .NET Core 3.0.0-preview-27122-01 (CoreCLR 4.6.27121.03, CoreFX 4.7.18.57103), 64bit RyuJIT


```
|        Method |     Mean |    Error |   StdDev |   Median |    Op/s | Gen 0/1k Op | Gen 1/1k Op | Gen 2/1k Op | Allocated Memory/Op |
|-------------- |---------:|---------:|---------:|---------:|--------:|------------:|------------:|------------:|--------------------:|
| ExecuteScalar | 125.7 us | 2.945 us | 8.683 us | 128.2 us | 7,953.8 |      0.2441 |           - |           - |             1.16 KB |

# After
``` ini

BenchmarkDotNet=v0.11.3, OS=ubuntu 18.10
Intel Xeon W-2133 CPU 3.60GHz, 1 CPU, 12 logical and 6 physical cores
.NET Core SDK=3.0.100-preview-009812
  [Host]     : .NET Core 3.0.0-preview-27122-01 (CoreCLR 4.6.27121.03, CoreFX 4.7.18.57103), 64bit RyuJIT
  DefaultJob : .NET Core 3.0.0-preview-27122-01 (CoreCLR 4.6.27121.03, CoreFX 4.7.18.57103), 64bit RyuJIT


```
|        Method |     Mean |    Error |   StdDev |   Median |    Op/s | Gen 0/1k Op | Gen 1/1k Op | Gen 2/1k Op | Allocated Memory/Op |
|-------------- |---------:|---------:|---------:|---------:|--------:|------------:|------------:|------------:|--------------------:|
| ExecuteScalar | 123.4 us | 2.555 us | 7.532 us | 125.5 us | 8,103.8 |      0.1221 |           - |           - |               824 B |