@YohDeadfall 2019-03-22
Would it be better to have a pool instead?

By the way, in the current implementation a type mapper can be alive for a long time after `ReloadTypes`.