@roji 2019-03-22
> Would it be better to have a pool instead?

You mean to avoid allocating a RowDescription for prepared statements? If so then that seems too much to me, it should be fine to pay a one-time cost when preparing (just like you pay the cost of server-side parsing and query planning). Or were you thinking of something else?

> By the way, in the current implementation a type mapper can be alive for a long time after ReloadTypes.

That's a very good point - it means that the cloned RowDescriptions held by prepared statements could reference a stale type mapper. So in theory, if types are reloaded and a type or mapping was removed or changed in some way, prepared statements would start failing. Can you think of a reasonable scenario where that 

However, we already have this problem, since FieldDescription already references the type mapper (and also the handler!), so I don't *think* this PR makes it worse. If that's true, then we can discuss it in a different issue - or am I mistaken?