@YohDeadfall 2019-03-23
I meant a type handler, made a typo (: Probably we should think about better design here. There is an `ObjectPool` in .NET extensions, but if my memory serves me I saw something like that in the Core too.