@roji 2019-03-25
@YohDeadfall sorry, I'm not understanding what you mean... :) What exactly do you propose that we pool? Type handlers already live as long as their physical connection (unless `ReloadTypes()` is called, which is rare). With this PR, the non-prepared RowDescription instance (and its fields) also live as long as its physical connection, and each `Prepare()` incurs one allocation as the RowDescription (and fields) are cloned.

Can you explain what you have in mind? If it's a separate improvement idea, maybe approve this and open a separate issue where we can discuss?