@roji 2019-04-01
> Don't have time to find a better way, so let's go ahead and merge.

OK, thanks - we can revisit later if you want.